requirements.txt
aiohttp==3.9.5
pandas==2.2.2
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1


python-telegram-bot>=20.0
flask>=2.0.0